DiscountedLineItemPriceForQuantity as interface #1053
