+++ commercetools-models/src/main/java/io/sphere/sdk/cartdiscounts/DiscountedLineItemPriceForQuantity.java
@@ -1,27 +1,14 @@
 package io.sphere.sdk.cartdiscounts;
 
-import com.fasterxml.jackson.annotation.JsonCreator;
-import io.sphere.sdk.models.Base;
+import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
 
-public final class DiscountedLineItemPriceForQuantity extends Base {
-    private final Long quantity;
-    private final DiscountedLineItemPrice discountedPrice;
-
-    @JsonCreator
-    private DiscountedLineItemPriceForQuantity(final DiscountedLineItemPrice discountedPrice, final Long quantity) {
-        this.discountedPrice = discountedPrice;
-        this.quantity = quantity;
+@JsonDeserialize(as = DiscountedLineItemPriceForQuantityImpl.class)
+public interface DiscountedLineItemPriceForQuantity {
+    static DiscountedLineItemPriceForQuantity of(final DiscountedLineItemPrice discountedLineItemPrice, final long quantity) {
+        return new DiscountedLineItemPriceForQuantityImpl(discountedLineItemPrice, quantity);
     }
 
-    public static DiscountedLineItemPriceForQuantity of(final DiscountedLineItemPrice discountedLineItemPrice, final long quantity) {
-        return new DiscountedLineItemPriceForQuantity(discountedLineItemPrice, quantity);
-    }
+    DiscountedLineItemPrice getDiscountedPrice();
 
-    public DiscountedLineItemPrice getDiscountedPrice() {
-        return discountedPrice;
-    }
-
-    public Long getQuantity() {
-        return quantity;
-    }
+    Long getQuantity();
 }
